+++ packages/jss-plugin-syntax-rule-value-function/src/index.d.ts
@@ -1,0 +1,3 @@
+import {Plugin} from 'jss'
+
+export default function jssPluginSyntaxRuleValueFunction(): Plugin